fix: address remaining PR #4 review comments
- Atomic schema hash write via os.replace to prevent partial-write corruption
- Replace bare get_connection()+try/finally with db._get_conn() context manager
- Document asyncio event-loop safety for _on_poller_update globals
- Use float() instead of int() for show_every_n_days cooldown calculation
- Delete dead _app_client() helper in test_suggestions.py
- Replace onMouseEnter/Leave style mutation with useState hover flags in SuggestionCard

+++ sidecar/suggestion_engine.py
@@ -107,7 +107,7 @@ def _in_cooldown(suggestion: dict, history: dict[str, dict]) -> bool:
         return False
     try:
         last_shown = datetime.fromisoformat(rec["shown_at"].replace("Z", "+00:00"))
-        next_ok = last_shown + timedelta(days=int(suggestion["show_every_n_days"]))
+        next_ok = last_shown + timedelta(days=float(suggestion["show_every_n_days"]))
         return datetime.now(tz=timezone.utc) < next_ok
     except (ValueError, TypeError):
         return False
